Added extractor trick. Refined 4-pool. Fixed atrocious bug in combat sim preventing melee units from attacking enemies in the fog. Fixed TrainContinuously (I hope) especially for Zerglings/Scourge. Applied more performance enhancements (?) in the form of UnitTracker.existsOurs and existsEnemy

+++ src/Micro/Matchups/MatchupAnalysis.scala
@@ -50,7 +50,7 @@ case class MatchupAnalysis(me: UnitInfo, conditions: MatchupConditions) {
   lazy val doomed                         : Boolean               = framesToLive <= framesOfEntanglement
   lazy val framesOfEntanglementPerThreat  : Map[UnitInfo, Double] = threats.map(threat => (threat, framesOfEntanglementWith(threat))).toMap
   lazy val framesOfEntanglement           : Double                = ByOption.max(framesOfEntanglementPerThreat.values).getOrElse(- Forever())
-  lazy val framesOfSafety                 : Double                = - With.latency.latencyFrames - With.reaction.agencyMax - ByOption.max(framesOfEntanglementPerThreat.values).getOrElse(- Forever())
+  lazy val framesOfSafety                 : Double                = - With.latency.latencyFrames - With.reaction.agencyMax - ByOption.max(framesOfEntanglementPerThreat.values).getOrElse(- Forever().toDouble)
   
   def dpfDealingDiffused  (target: UnitInfo): Double = me.dpfOnNextHitAgainst(target) / Math.max(1.0, targetsInRange.size)
