+++ src/Utilities/Forever.scala
@@ -1,5 +1,5 @@
 package Utilities
 
 object Forever {
-  def apply(): Double = 24 * 60 * 60
+  def apply(): Int = 24 * 60 * 60
 }
